+++ src/simulation/body/particle/ChargedParticle.java
@@ -59,7 +59,7 @@ public class ChargedParticle extends Particle implements Gravity {
   @Override
   public short getNearRadius() {
     return (short) (
-      Settings.get(Settings.Constants.MAX_RADIUS) + (float) strength / 3
+      Settings.get(Settings.Constants.MAX_RADIUS) + strength / 3f
     );
   }
